backend/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2

# Deep learning — CPU wheel
torch==2.2.2
torchvision==0.17.2

# Model architecture (phải khớp với lúc train)
segmentation-models-pytorch==0.3.3

# Image augmentation / preprocessing
# Pin opencv version cụ thể để tránh pip thử nhiều version
opencv-python-headless==4.9.0.80
albumentations==1.3.1

# Dependency của pretrainedmodels (dùng six.moves.urllib)
six
